agent_1/pyproject.toml
[project]
dependencies = [
  "langchain>=1.2.15",
  "python-dotenv>=1.2.2",
  "transformers>=5.7.0",
]
description = "An agent"
name = "agent-1"
readme = "README.md"
requires-python = ">=3.14"
version = "0.1.0"

[dependency-groups]
dev = [
  "black>=26.3.1",
  "pyright>=1.1.409",
  "ruff>=0.15.12",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I"] # import order

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pyright]
reportMissingImports = true
reportUnusedVariable = true
typeCheckingMode = "strict"
